+++ src/index.js
@@ -67,6 +67,7 @@ export {
 export {
   Node,
   isNode,
+  isNodeLike,
   isNodeObj,
   isWrittenNode,
   isSameNode,